Extract DynamoDB boilerplate from VersionTableTestCase
